[example] move GQLViewer into its own file
decorator()
export class

seems to lead to issues in TypeScript

googling points to https://github.com/microsoft/TypeScript/issues/53332
